up to abfab98

+++ src/build-linux.sh
@@ -1,7 +1,22 @@
 #!/bin/sh
+# ABOUT
+# Build shell script for Artisan Linux builds
+#
+# LICENSE
+# This program or module is free software: you can redistribute it and/or
+# modify it under the terms of the GNU General Public License as published
+# by the Free Software Foundation, either version 2 of the License, or
+# version 3 of the License, or (at your option) any later versison. It is
+# provided for educational purposes and is distributed in the hope that
+# it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+# the GNU General Public License for more details.
+#
+# AUTHOR
+# Dave Baxter, Marko Luther 2023
 
-#set -ex
-set -e  # reduced logging
+#set -ex # extended logging
+set -e  # exit as soon as any line in the script fails
 
 
 export LD_LIBRARY_PATH=$LD_LIBTRARY_PATH:/usr/local/lib
@@ -10,18 +25,24 @@ export PATH=$PATH:$HOME/.local/bin
 if [ ! -z $APPVEYOR ]; then
     # Appveyor environment
     echo "NOTICE: Appveyor build"
-    #PYTHON_PATH passed from appveyor.yml
-    #export PYTHON_PATH=/home/appveyor/venv3.10.6/lib/python3.10/site-packages
-    export QT_PATH=$PYTHON_PATH/PyQt6/Qt6
-elif [ -d /usr/lib/python3/dist-packages/PyQt5 ]; then
-    # ARM builds
-    export PYTHON_PATH=`python3 -m site --user-site`
-    export QT_PATH=/usr/share/qt5
+elif [ -d /usr/lib/python3/dist-packages/PyQt6 ]; then
+    # ARM RPi bookworm builds (assumes requirements.txt installed in a user local venv
+    export PYTHON_PATH=`python3 -c 'import sysconfig; print(sysconfig.get_paths()["purelib"])'`
+    export PYTHONSITEPKGS=`python3 -c 'import sysconfig; print(sysconfig.get_paths()["purelib"])'`
+    export QT_PATH=/usr/share/qt6
+    export QT_SRC_PATH=/usr/lib/qt6
+#    export PYUIC=`python3 -m PyQt6.uic.pyuic`
+    export PYLUPDATE=./pylupdate6pro.py
+#    export PYLUPDATE=/usr/bin/pylupdate6
 else
     # Other builds
     export PYTHON_PATH=`python3 -m site --user-site`
     export QT_PATH=$PYTHON_PATH/PyQt5/Qt
 fi
+
+echo "************* build derived files **************"
+./build-derived.sh linux  #generate the derived files
+if [ $? -ne 0 ]; then echo "Failed in build-derived.sh"; exit $?; else (echo "** Finished build-derived.sh"); fi
 
 rm -rf build
 rm -rf dist
@@ -43,25 +64,11 @@ rm -rf dist/artisan.d
 
 # copy translations
 mkdir dist/translations
-
-
-for lan in ar, da, de, en, el, es, fa, fi, fr, gd, he, hu, id, it, ja, ko, lv, nl, no, pl, pt_BR, pt, sk, sv, th, tr, uk, vi, zh_CN, zh_TW; do
-     QTBASE_FILE=$QT_PATH/translations/qtbase_${lan}.qm
-     QT_FILE=$QT_PATH/translations/qt_${lan}.qm
-     if [ -e ${QTBASE_FILE} ]
-          then cp ${QTBASE_FILE} dist/translations
-     fi
-     if [ -e ${QT_FILE} ]
-          then cp ${QT_FILE} dist/translations
-     fi
-done
-
 cp translations/*.qm dist/translations
 
 # copy data (mpl-data is now copied to matplotlib/mpl-data by pyinstaller)
 #cp -R $PYTHON_PATH/matplotlib/mpl-data dist
-#rm -rf dist/mpl-data/sample_data
-rm -rf dist/matplotlib/sample_data
+rm -rf dist/_internal/matplotlib/sample_data
 
 # copy file icon and other includes
 cp artisan-alog.xml dist
@@ -79,17 +86,34 @@ cp includes/SourceHanSansJP-Regular.otf dist
 cp includes/SourceHanSansKR-Regular.otf dist
 cp includes/SourceHanSansTW-Regular.otf dist
 cp includes/dijkstra.ttf dist
+cp includes/ComicNeue-Regular.ttf dist
+cp includes/xkcd-script.ttf dist
 cp includes/alarmclock.eot dist
 cp includes/alarmclock.svg dist
 cp includes/alarmclock.ttf dist
 cp includes/alarmclock.woff dist
+cp includes/roboto-300.woff2 dist
+cp includes/roboto-600.woff2 dist
+cp includes/roboto-regular.woff2 dist
 cp includes/artisan.tpl dist
+cp includes/scale_widget.tpl dist
+cp includes/fitty_patched.js dist
 cp includes/bigtext.js dist
 cp includes/sorttable.js dist
 cp includes/report-template.htm dist
 cp includes/roast-template.htm dist
 cp includes/ranking-template.htm dist
 cp includes/jquery-1.11.1.min.js dist
+cp includes/android-chrome-192x192.png dist
+cp includes/android-chrome-512x512.png dist
+cp includes/apple-touch-icon.png dist
+cp includes/browserconfig.xml dist
+cp includes/favicon-16x16.png dist
+cp includes/favicon-32x32.png dist
+cp includes/favicon.ico dist
+cp includes/mstile-150x150.png dist
+cp includes/safari-pinned-tab.svg dist
+cp includes/site.webmanifest dist
 cp includes/logging.yaml dist
 cp -R icons dist
 cp -R Wheels dist
@@ -108,44 +132,97 @@ mkdir dist/Icons
 find includes/Icons -name '.*.aset' -exec rm -r {} \;
 cp -R includes/Icons/* dist/Icons
 
-mkdir dist/yoctopuce
-mkdir dist/yoctopuce/cdll
-cp $PYTHON_PATH/yoctopuce/cdll/*64.so dist/yoctopuce/cdll
 
-cp /usr/lib/libsnap7.so dist
+# remove unused Qt modules
 
-cp README.txt dist
-cp ../LICENSE dist/LICENSE.txt
+keep_qt_modules="libQt6Bluetooth libQt6Concurrent libQt6Core libQt6DBus libQt6Gui libQt6Network
+ libQt6OpenGL libQt6Positioning libQt6PrintSupport libQt6Qml libQt6QmlModels libQt6Quick libQt6QuickWidgets
+ libQt6Svg libQt6WaylandClient libQt6WaylandEglClientHwIntegration libQt6WebChannel libQt6WebEngineCore
+ libQt6WebEngineWidgets libQt6Widgets libQt6WlShellIntegration libQt6XcbQpa"
 
-# remove automatically collected PyQt6 libs that are not used to save space
-rm -f dist/libQt6Multimedia*.*
-rm -f dist/libQt6Quick3D*.*
-rm -f dist/libQt6QuickC*.*
-rm -f dist/libQt6QuickD*.*
-rm -f dist/libQt6QuickD*.*
-rm -f dist/libQt6QuickL*.*
-rm -f dist/libQt6QuickP*.*
-rm -f dist/libQt6QuickT*.*
-rm -f dist/libQt6QuickS*.*
-rm -f dist/libQt6RemoteObjects*.*
-rm -f dist/libQt6Sensors*.*
-rm -f dist/libQt6SerialPort*.*
-rm -f dist/libQt6ShaderTools*.*
-rm -f dist/libQt6Sql*.*
-rm -f dist/libQt6Test*.*
-rm -rf dist/PyQt6/Qt6/translations
-rm -rf dist/PyQt6/Qt6/qml/QtQuick3D
+for qtlib in $(find dist/_internal/PyQt6/Qt6/lib -type f -name "libQt6*.so.*"); do
+    qtlib_filename="${qtlib##*/}"
+    match=0
+    for item in ${keep_qt_modules}; do
+        if [ ${qtlib_filename} = "${item}.so.6" ]; then
+            match=1
+            break
+        fi
+    done
+    if [ $match = 0 ]; then
+        rm -f ${qtlib}
+    fi
+done
+
+# remove Qt5 libs
+rm -rf dist/_internal/libQt5*
+rm -rf dist/_internal/PyQt5
+
+
+# remove unused Qt plugins
+
+qt_imageformats="libqwebp libqtiff libqgif libqwbmp libqtga"
+for x in ${qt_imageformats}; do
+    rm -rf "dist/_internal/PyQt6/Qt6/plugins/imageformats/${x}.so"
+done
+
+
+SUPPORTED_LANGUAGES="ar cs da de el en es fa fi fr gd he hu id it ja ko lv nl no pl pt_BR pt sk sv th tr uk vi zh_CN zh_TW"
+
+# remove unused Qt translations
+
+# the following produces a (harmless) warning log entry on generating PDF reports as locales cannot be found
+rm -rf dist/_internal/PyQt6/Qt6/translations/qtwebengine_locales
+
+for qttrans in $(find dist/_internal/PyQt6/Qt6/translations -type f -name "*.qm"); do
+    qttrans_filename="${qttrans##*/}"
+    match=0
+    for lang in ${SUPPORTED_LANGUAGES}; do
+        if [ ${qttrans_filename} = "qtbase_${lang}.qm" ] || [ ${qttrans_filename} = "qt_${lang}.qm" ] ; then
+            match=1
+            break
+        fi
+    done
+    if [ $match = 0 ]; then
+        rm -f ${qttrans}
+    fi
+done
+
+
+# remove unused QML files
+
+rm -rf dist/_internal/PyQt6/Qt6/qml
+echo Qt QML files removed
+
+# remove unused babel translations
+
+for babeltrans in $(find dist/_internal/babel/locale-data -type f -name "*.dat"); do
+    babeltrans_filename="${babeltrans##*/}"
+    match=0
+    for lang in ${SUPPORTED_LANGUAGES}; do
+        if [ ${babeltrans_filename} = "${lang}.dat" ] || [ ${babeltrans_filename} = "root.dat" ] || [ ${babeltrans_filename} = "zh_Hans_CN.dat" ] || [ ${babeltrans_filename} = "zh_Hant_TW.dat" ] || [ ${babeltrans_filename} == "zh.dat" ] ; then
+            match=1
+            break
+        fi
+    done
+    if [ $match = 0 ]; then
+        rm -f ${babeltrans}
+    fi
+done
+
+# remove matplotlib sample data
+rm -rf dist/_internal/matplotlib/mpl-data/sample_data
 
 # remove automatically collected libs that might break things on some installations (eg. Ubuntu 16.04)
 # so it is better to rely on the system installed once
 # see https://github.com/gridsync/gridsync/issues/47 and https://github.com/gridsync/gridsync/issues/43
 #   and https://askubuntu.com/questions/575505/glibcxx-3-4-20-not-found-how-to-fix-this-error
-rm -f dist/libdrm.so.2 # https://github.com/gridsync/gridsync/issues/47
-rm -f dist/libX11.so.6 # https://github.com/gridsync/gridsync/issues/43
-rm -f dist/libstdc++.so.6 # https://github.com/gridsync/gridsync/issues/189
+rm -f dist/_internal/libdrm.so.2 # https://github.com/gridsync/gridsync/issues/47
+rm -f dist/_internal/libX11.so.6 # https://github.com/gridsync/gridsync/issues/43
+rm -f dist/_internal/libstdc++.so.6 # https://github.com/gridsync/gridsync/issues/189
 # the following libs might not need to be removed
-rm -f dist/libgio-2.0.so.0
-rm -f dist/libz.so.1 # removing this lib seems to break the build on some RPi Buster version
-rm -f dist/libglib-2.0.so.0 # removed for v1.6 and later
+rm -f dist/_internal/libgio-2.0.so.0
+rm -f dist/_internal/libz.so.1 # removing this lib seems to break the build on some RPi Buster version
+rm -f dist/_internal/libglib-2.0.so.0 # removed for v1.6 and later
 
 rm -f libusb-1.0.so.0
